@@ -27,64 +28,66 @@
     "test-storybook": "test-storybook"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.0",
-    "@babel/preset-env": "^7.16.4",
-    "@babel/preset-react": "^7.16.0",
-    "@babel/preset-typescript": "^7.16.0",
-    "@mdx-js/react": "^2.1.2",
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-image": "^3.0.1",
-    "@rollup/plugin-node-resolve": "^13.0.6",
-    "@rollup/plugin-typescript": "8.3.3",
-    "@storybook/addon-actions": "^7.0.6",
-    "@storybook/addon-essentials": "^7.0.6",
-    "@storybook/addon-interactions": "^7.0.6",
-    "@storybook/addon-links": "^7.0.6",
-    "@storybook/addon-mdx-gfm": "^7.0.6",
-    "@storybook/jest": "^0.1.0",
-    "@storybook/manager-api": "^7.5.3",
+    "@babel/core": "^7.26.0",
+    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-react": "^7.26.3",
+    "@babel/preset-typescript": "^7.26.0",
+    "@chromatic-com/storybook": "^3",
+    "@mdx-js/react": "^3.1.0",
+    "@rollup/plugin-commonjs": "^28.0.2",
+    "@rollup/plugin-image": "^3.0.3",
+    "@rollup/plugin-node-resolve": "^16.0.0",
+    "@rollup/plugin-typescript": "12.1.2",
+    "@storybook/addon-actions": "^8.4.7",
+    "@storybook/addon-essentials": "^8.4.7",
+    "@storybook/addon-interactions": "^8.4.7",
+    "@storybook/addon-links": "^8.4.7",
+    "@storybook/addon-webpack5-compiler-babel": "^3.0.3",
+    "@storybook/manager-api": "^8.4.7",
     "@storybook/preset-scss": "^1.0.3",
-    "@storybook/react": "^7.0.6",
-    "@storybook/react-webpack5": "^7.0.6",
-    "@storybook/test-runner": "^0.10.0",
-    "@storybook/testing-library": "^0.1.0",
-    "@storybook/theming": "^7.5.3",
-    "@svgr/rollup": "^6.5.1",
-    "@svgr/webpack": "^6.5.1",
-    "@testing-library/react": "^13.3.0",
-    "@types/dompurify": "^2.4.0",
-    "@types/jest": "^28.1.6",
-    "@types/react": "^18.0.15",
-    "babel-jest": "^27.3.1",
-    "babel-loader": "^8.2.3",
-    "css-loader": "^6.5.1",
-    "dompurify": "^3.2.0",
+    "@storybook/react": "^8.4.7",
+    "@storybook/react-webpack5": "^8.4.7",
+    "@storybook/test": "^8.4.7",
+    "@storybook/test-runner": "^0.20.1",
+    "@storybook/theming": "^8.4.7",
+    "@storybook/types": "^8.4.7",
+    "@svgr/rollup": "^8.1.0",
+    "@svgr/webpack": "^8.1.0",
+    "@testing-library/react": "^16.1.0",
+    "@types/dompurify": "^3.2.0",
+    "@types/jest": "^29.5.14",
+    "@types/react": "^19.0.2",
+    "babel-jest": "^29.7.0",
+    "babel-loader": "^9.2.1",
+    "css-loader": "^7.1.2",
+    "dompurify": "^3.2.3",
     "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.0",
+    "html-webpack-plugin": "^5.6.3",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^28.1.3",
-    "jest-environment-jsdom": "^29.3.1",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "jest-svg-transformer": "^1.0.0",
     "pullstate": "^1.25.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-tooltip": "^5.28.0",
-    "rollup": "^2.79.1",
-    "rollup-plugin-dts": "^4.0.1",
+    "rollup": "^4.28.1",
+    "rollup-plugin-dts": "^6.1.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-scss": "^4.0.0",
+    "rollup-plugin-scss": "^4.0.1",
     "rollup-plugin-terser": "^7.0.2",
-    "sass": "^1.43.5",
-    "sass-loader": "^12.3.0",
-    "storybook": "^7.0.6",
-    "style-loader": "^3.3.1",
-    "tsconfig-paths-webpack-plugin": "^4.0.0",
-    "typescript": "^4.7.4"
+    "sass": "^1.83.0",
+    "sass-loader": "^16.0.4",
+    "storybook": "^8.4.7",
+    "style-loader": "^4.0.0",
+    "tsconfig-paths-webpack-plugin": "^4.2.0",
+    "typescript": "^5.7.2",
+    "isomorphic-dompurify": "^2.19.0"
   },
   "peerDependencies": {
-    "isomorphic-dompurify": "^2.16.0",
-    "react": "^18.3.1",
+    "isomorphic-dompurify": "^2.19.0",
+    "react": "^19.0.0",
     "react-tooltip": "^5.28.0"
   }
 }
