@@ -1,13 +1,13 @@
-import React from 'react';
-import {Meta, StoryFn} from '@storybook/react';
+import React from "react";
+import {Meta, StoryFn} from "@storybook/react";
 import {ReactComponent as AlarmBellIcon} from "../../icons/functional-icons/allowance-no-talking.svg";
 
-import Badge from './Badge';
+import Badge from "./Badge";
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-    title: 'ReactComponentLibrary/Badge',
-    component: Badge,
+  title: "ReactComponentLibrary/Badge",
+  component: Badge,
 } as Meta<typeof Badge>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
@@ -16,12 +16,11 @@ const Template: StoryFn<typeof Badge> = (args) => <Badge {...args} />;
 export const BadgeIconTemplate = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 BadgeIconTemplate.args = {
-    label: "Badge    text",
-    children: <AlarmBellIcon/>
+  label: "Badge    text",
+  children: <AlarmBellIcon />,
 };
 
 export const BadgeTemplate = Template.bind({});
 BadgeTemplate.args = {
-    label: "Badge text"
+  label: "Badge text",
 };
-
