@@ -1,4 +1,4 @@
-import React, {useState} from "react";
+import React, {JSX, useState} from "react";
 import "./ContentCard.scss";
 import {stopEvent} from "../../common/utils";
 
